% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read}
\alias{read}
\title{Read a file}
\usage{
read(file, skip = 0, na = "")
}
\arguments{
\item{file}{A path to a file.}

\item{skip}{Number of lines to skip before reading data.}

\item{na}{Character vector of strings to interpret as missing values.}
}
\description{
Read a dataset file.
}
