% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskymsemi-package.R
\docType{package}
\name{tskymsemi-package}
\alias{tskymsemi}
\alias{tskymsemi-package}
\title{tskymsemi: Convenient functions for Tsukiyama Seminar}
\description{
This package includes convenient functions for Tsukiyama Seminar. e.g. Creating a descriptive statistics table.
}
\author{
\strong{Maintainer}: First Last \email{0122atk@gmail.com}

}
\keyword{internal}
