% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write}
\alias{write}
\title{Write a file}
\usage{
write(data, file)
}
\arguments{
\item{data}{A data frame or tibble to write to disk.}

\item{file}{File to write to.}
}
\description{
Write a dataset file.
}
